Improve PlayStats: weight of 0 is now impossible
Primary motivation is to avoid weight=0 when player gets hand correct
on their first try. Weight of zero means they'll never see it again.
All weights of zero is an error in WeightedDistribution.

The change also means that getting 1 of 2 hands correct is no longer
the same weight as getting 10 of 20 hands correct.