Change PlayStats to use u32 instead of u16
A future change will try to make certain hands essentially impossible
to be generated by giving them PlayStats with seen == correct at the
uint's max value minus 1. When using u16, this would result in a
weight of 1.52 x 10^-5, which is not thaaaaaaaat unlikely.

    1 - ((2^16)-1)/(2^16)

But when using u32, this is reduced to 2.32 x 10^-10, which should be
close enough to zero.