Make web button enum with hit, stand, etc. ...
Move "did the player choose the correct move" logic into its own function.
Write tests for it.

This is motivated by a bug where the Double button would be pressed, it
would be changed to DoubleElseHit (as an abitrary programming close), and
the cell would have DoubleElseStand as the correct move. Thus the player
would get told that Double is wrong, Double(Stand) is correct.