Add (de)serialization util funcs for Table<PlayStats> ...
So that wasm web code can call it. The point is for something that
can be stored in a DB easily with minimal overhead while also not
using up too much dev time to minimize size.

In the expected worst case (it could be worse, technically), the string
representation of a Table<PlayStats> will take up 360*6-1 bytes:
"99/99," for each of the 360 cells in the table, minus 1 for removing
the last trailing comma. This is 2159 bytes.

Alternatives:

- Just take the json: ~23 KB. Too big.
- Take the json, and compress it: ~970 bytes (with xz). Quite good, but
leaves a bad taste in my mouth. Storing compressed json in a DB?! Who
does that?

I've convinced myself that this custom serialized string is less weird.